Added `CommonStorage` module
