+++ chain-factory/src/factory.rs
@@ -31,7 +31,9 @@ pub enum ScArray {
 }
 
 #[multiversx_sc::module]
-pub trait FactoryModule: only_admin::OnlyAdminModule {
+pub trait FactoryModule:
+    only_admin::OnlyAdminModule + crate::common::storage::CommonStorage
+{
     // TODO: Check if contract was already deployed
     #[payable("EGLD")]
     #[endpoint(deploySovereignChainConfigContract)]
@@ -272,49 +274,4 @@ pub trait FactoryModule: only_admin::OnlyAdminModule {
             "The number of validator BLS Keys is not correct"
         );
     }
-
-    #[view(getContractsMap)]
-    #[storage_mapper("contractsMap")]
-    fn contracts_map(&self, contract_name: ScArray) -> SingleValueMapper<ManagedAddress>;
-
-    #[view(getDeployCost)]
-    #[storage_mapper("deployCost")]
-    fn deploy_cost(&self) -> SingleValueMapper<BigUint>;
-
-    #[storage_mapper("chainConfigTemplate")]
-    fn chain_config_template(&self) -> SingleValueMapper<ManagedAddress>;
-
-    #[storage_mapper("headerVerifierTemplate")]
-    fn header_verifier_template(&self) -> SingleValueMapper<ManagedAddress>;
-
-    #[storage_mapper("crossChainOperationsTemplate")]
-    fn cross_chain_operations_template(&self) -> SingleValueMapper<ManagedAddress>;
-
-    #[storage_mapper("feeMarketTemplate")]
-    fn fee_market_template(&self) -> SingleValueMapper<ManagedAddress>;
-
-    #[storage_mapper("tokenHandlerTemplate")]
-    fn token_handler_template(&self) -> SingleValueMapper<ManagedAddress>;
-
-    #[storage_mapper("allDeployedContracts")]
-    fn all_deployed_contracts(
-        &self,
-        chain_id: ManagedBuffer,
-    ) -> UnorderedSetMapper<ContractMapArgs<Self::Api>>;
-
-    #[storage_mapper_from_address("minValidators")]
-    fn external_min_validators(
-        &self,
-        sc_address: ManagedAddress,
-    ) -> SingleValueMapper<BigUint<Self::Api>, ManagedAddress>;
-
-    #[storage_mapper_from_address("maxValidators")]
-    fn external_max_validators(
-        &self,
-        sc_address: ManagedAddress,
-    ) -> SingleValueMapper<BigUint<Self::Api>, ManagedAddress>;
-
-    #[view(getAllChainIds)]
-    #[storage_mapper("allChainIds")]
-    fn chain_ids(&self) -> UnorderedSetMapper<ManagedBuffer>;
 }
